Update lambda_function.py
Adds conversation starter functionality.

The function now checks if the input_text.assistant_input is set to anything apart from 'none' or empty and then starts a conversation based on that prompt. 

If the input_text.assistant_input is set to literally 'none' or nothing then the assistant will start without a prompt as set in the language files.